feat(visorconfig): extract schema leaf packages — V1 now WASM-clean
Three-layer separation per the architectural plan: every config-shaped
type V1 embeds gets a schema sub-package (Layer 1, pure data + JSON
methods, stdlib + cipher only) that the original operational package
(Layer 2) re-exports via type alias. V1 (Layer 3) references only
Layer 1 types as its fields. End result: pkg/visor/visorconfig
compiles under GOOS=js/GOARCH=wasm — the in-browser config-generator
(apt-repo install page) can now construct + emit live visor configs
using the canonical V1 type instead of a parallel definition that
would drift.

Schema sub-packages added:

  pkg/dmsgc/spec               Deployment, DmsgConfig (pure JSON
                                marshalers + PKFromDmsgURL helper);
                                pkg/dmsgc keeps the dmsg.Client
                                constructor + lanPriorityDisc.

  pkg/transport/network/spec   STCPConfig (PKTable + listening
                                address); pkg/transport/network
                                keeps the STCP client / dialer /
                                listener.

  pkg/transport/spec           PersistentTransports (PK + NetType);
                                pkg/transport keeps the Manager
                                and its reconciliation loop.

  pkg/app/appserver/spec       AppConfig (the on-disk per-app entry
                                that V1.Launcher.Apps embeds);
                                pkg/app/appserver keeps the
                                AppLauncher, server logic, etc.

visorconfig changes:

  v1.go                Imports switched to spec packages; field
                        types now reference the leaves directly.
                        Now pure-data + pure-method only —
                        compiles under js.

  v1_runtime.go (NEW)  Build-tag-gated to !js. Holds every method
                        that takes *launcher.AppLauncher (Update*,
                        DeleteAppConfig, AddAppConfig, etc.) and
                        the helpers they call (updateStringArg /
                        updateBoolArg / deleteAppArg / updateEnvEntry).
                        Native builds see the full V1 surface
                        unchanged; WASM builds see V1 minus those
                        12 methods, which is exactly what an
                        in-browser config-generator needs.

  args.go              appsList re-typed from []appserver.AppConfig
                        to []appspec.AppConfig (identical via
                        alias, but spec import is WASM-clean).

  config.go            MakeBaseConfig now uses dmsgspec.DmsgConfig
                        + tnspec.STCPConfig. The one host-specific
                        call (dmsgpty.DefaultCLIAddr() — pulls in
                        pty package which has no js support)
                        replaced by a local defaultDmsgPtyCLIAddr()
                        helper returning the unix temp-socket path
                        literal. cli-side config gen (cmd/skywire-cli)
                        still calls dmsgpty.DefaultCLIAddr() in its
                        native build for proper windows path
                        resolution.

  values.go            DmsgHTTPPort = uint16(80) literal instead of
                        importing pkg/dmsg/dmsg's
                        DefaultDmsgHTTPPort. The git-describe
                        version fallback moved to a new
                        values_native.go (!js) — js builds skip the
                        shell-out and return buildinfo.Version()
                        unchanged.

  values_js.go (NEW)   js stubs for UserConfig, SystemSurvey,
                        IsRoot, Survey. Browser has no host
                        filesystem / hardware inventory; the
                        functions exist for type-resolution under
                        GOOS=js and return zero values.

  values_native.go (NEW)
                        Native (!js) helpers extracted from values.go:
                        currently just resolveVersion's git-describe
                        fallback.

pkg/skyenv/skyenv_js.go    PackageConfig() stub added (returns zero
                            PkgConfig). Needed for visorconfig's
                            values.go to compile under js.

Other callsites — pkg/dmsgc.DmsgConfig, network.STCPConfig,
transport.PersistentTransports, appserver.AppConfig — all keep
working unchanged via type aliases. Native tests pass across the
touched packages:

  pkg/visor/visorconfig
  pkg/dmsgc
  pkg/transport, pkg/transport/network
  pkg/app/appserver

Verified compile under GOOS=js GOARCH=wasm — 8.1 MB blob, no
external deps required for the apt-repo install page's WASM to
reference visorconfig.V1 directly.

Followup that builds on this: pkg/skywireconfig/genvisor will
provide a Generate(opts) -> *V1 entry point that composes
deployment.Prod + operator opts using only the now-WASM-clean V1
type. Then apt-repo/cmd/wasm can emit a downloadable
/opt/skywire/skywire.json from the install-page form.
