+++ pkg/app/appserver/app_state.go
@@ -1,7 +1,9 @@
 // Package appserver pkg/app/appserver/app_state.go
 package appserver
 
-import "github.com/skycoin/skywire/pkg/routing"
+import (
+	"github.com/skycoin/skywire/pkg/app/appserver/spec"
+)
 
 // AppStatus defines running status of an App.
 type AppStatus int
@@ -44,34 +46,11 @@ func (s AppStatus) String() string {
 // definition — credentials can't be dropped per-goroutine without
 // breaking the runtime's thread-pinning assumptions, and the visor
 // logs a warning if any of these fields are set on an internal app.
-type AppConfig struct {
-	Name      string       `json:"name"`
-	Binary    string       `json:"binary,omitempty"`
-	Args      []string     `json:"args,omitempty"`
-	AutoStart bool         `json:"auto_start"`
-	Port      routing.Port `json:"port"`
-	// User overrides the UID the spawned process runs as (POSIX
-	// setuid before exec). Requires the visor to have the
-	// privileges to switch users (i.e. running as root, or with
-	// CAP_SETUID). Empty = inherit visor UID.
-	User string `json:"user,omitempty"`
-	// Group overrides the GID. Empty = inherit visor GID, or the
-	// primary group of User when User is set.
-	Group string `json:"group,omitempty"`
-	// WorkDir overrides the proc's working directory. Empty =
-	// launcher's per-app default (lc.LocalPath/<app-name>).
-	WorkDir string `json:"work_dir,omitempty"`
-	// Env adds key=value pairs to the spawned process's
-	// environment. Useful for HOME=/home/<user> when User is set.
-	Env []string `json:"env,omitempty"`
-	// LauncherMode persists the operator's launcher-mode preference
-	// for this app: "internal" runs the app inside the visor process,
-	// "external" spawns it as a child with optional User/Group/etc.
-	// Empty = use whatever the visor's StartApp default chooses
-	// (today: external when Binary is set, internal otherwise).
-	// Honored by StartApp; StartAppWithMode still wins per-call.
-	LauncherMode string `json:"launcher_mode,omitempty"`
-}
+// AppConfig is the wire-format type for an app's launcher entry.
+// Aliased from pkg/app/appserver/spec so existing callers writing
+// `appserver.AppConfig{...}` keep compiling, while the canonical
+// (WASM-clean) definition lives in the spec leaf package.
+type AppConfig = spec.AppConfig
 
 // AppState defines state parameters for a registered App.
 type AppState struct {
