+++ pkg/visor/visorconfig/args.go
@@ -16,14 +16,17 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/skycoin/skywire/pkg/app/appserver"
+	appspec "github.com/skycoin/skywire/pkg/app/appserver/spec"
 	"github.com/skycoin/skywire/pkg/routing"
 )
 
-// appsList wraps []appserver.AppConfig so the JSON I/O layer can
+// appsList wraps []appspec.AppConfig so the JSON I/O layer can
 // rewrite Args between string and []string without touching the
-// shared appserver type.
-type appsList []appserver.AppConfig
+// shared appserver type. Uses the WASM-clean spec subpackage so
+// pkg/visor/visorconfig stays compilable under GOOS=js — the type
+// is identical to appspec.AppConfig (alias) so callers that read
+// the on-disk config through this package keep working unchanged.
+type appsList []appspec.AppConfig
 
 // MarshalJSON emits each AppConfig's Args as a shell-quoted string.
 func (a appsList) MarshalJSON() ([]byte, error) {
@@ -41,7 +44,7 @@ func (a *appsList) UnmarshalJSON(data []byte) error {
 	if err := json.Unmarshal(data, &raw); err != nil {
 		return err
 	}
-	out := make([]appserver.AppConfig, len(raw))
+	out := make([]appspec.AppConfig, len(raw))
 	for i, r := range raw {
 		cfg, err := unmarshalAppConfig(r)
 		if err != nil {
@@ -53,7 +56,7 @@ func (a *appsList) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
-// appConfigOnDisk mirrors appserver.AppConfig with Args as a string.
+// appConfigOnDisk mirrors appspec.AppConfig with Args as a string.
 // Only used at the JSON boundary.
 type appConfigOnDisk struct {
 	Name         string       `json:"name"`
@@ -68,7 +71,7 @@ type appConfigOnDisk struct {
 	LauncherMode string       `json:"launcher_mode,omitempty"`
 }
 
-func toOnDisk(c appserver.AppConfig) appConfigOnDisk {
+func toOnDisk(c appspec.AppConfig) appConfigOnDisk {
 	return appConfigOnDisk{
 		Name:         c.Name,
 		Binary:       c.Binary,
@@ -83,15 +86,15 @@ func toOnDisk(c appserver.AppConfig) appConfigOnDisk {
 	}
 }
 
-func unmarshalAppConfig(raw json.RawMessage) (appserver.AppConfig, error) {
+func unmarshalAppConfig(raw json.RawMessage) (appspec.AppConfig, error) {
 	// Shadow AppConfig.Args with a RawMessage at the outer scope so
 	// we can dispatch on its JSON shape (string vs array).
 	var pre struct {
-		appserver.AppConfig
+		appspec.AppConfig
 		Args json.RawMessage `json:"args,omitempty"`
 	}
 	if err := json.Unmarshal(raw, &pre); err != nil {
-		return appserver.AppConfig{}, err
+		return appspec.AppConfig{}, err
 	}
 	cfg := pre.AppConfig
 	if len(pre.Args) == 0 {
@@ -104,13 +107,13 @@ func unmarshalAppConfig(raw json.RawMessage) (appserver.AppConfig, error) {
 	if err := json.Unmarshal(pre.Args, &s); err == nil {
 		parsed, perr := splitArgs(s)
 		if perr != nil {
-			return appserver.AppConfig{}, fmt.Errorf("parsing args string: %w", perr)
+			return appspec.AppConfig{}, fmt.Errorf("parsing args string: %w", perr)
 		}
 		cfg.Args = parsed
 		return cfg, nil
 	}
 	if err := json.Unmarshal(pre.Args, &cfg.Args); err != nil {
-		return appserver.AppConfig{}, fmt.Errorf("args must be string or array: %w", err)
+		return appspec.AppConfig{}, fmt.Errorf("args must be string or array: %w", err)
 	}
 	return cfg, nil
 }
